LayoutTests/imported/w3c/web-platform-tests/html/syntax/parsing/adoption_agency_check_the_end_tag_name-expected.txt


FAIL The algorithm should not reparent properly nested tags assert_equals: expected "<code some-attribute=\"\"><div><code><code><code><code></code></code></code></code></div></code>" but got "<code some-attribute=\"\"></code><div><code some-attribute=\"\"><code><code><code><code></code></code></code></code></code></div>"
